refactor: remove extra volume mounts from install options
The extra volume mounts field was not actually used as an
install option; it is only needed during cluster provisioning.
Removing it from the install options allows us to move the
initialization of install options further down the code path,
after cluster creation.

This enables actions such as querying Helm or the cluster when
constructing Helm values, etc.